test: corrected tests, new shared logic, fixed example
