+++ tests/arithmetic.rs
@@ -1,6 +1,4 @@
-mod shared;
-use shared::*;
-use silt_lua::ExVal;
+use silt_lua::{test_number, valeq, ExVal,simple};
 
 test_number!(add_integers, "return 5 + 3", 8.0);
 test_number!(subtract_integers, "return 10 - 4", 6.0);
