+++ tests/functions.rs
@@ -1,6 +1,4 @@
-mod shared;
-use shared::*;
-use silt_lua::ExVal;
+use silt_lua::{ExVal,valeq,simple};
 
 #[test]
 fn simple_function() {
