+++ tests/variables.rs
@@ -1,6 +1,5 @@
-mod shared;
-use shared::*;
-use silt_lua::ExVal;
+use crate::{};
+use silt_lua::{ExVal,simple, test_number, test_string, valeq, test_nil, test_bool};
 
 test_number!(local_variable, "local x = 42; return x", 42.0);
 test_string!(
